+++ classes/sitemapentry.php
@@ -1,5 +1,5 @@
 <?php
-namespace Grav\Plugin;
+namespace Grav\Plugin\Sitemap;
 
 class SitemapEntry
 {
@@ -8,4 +8,22 @@ class SitemapEntry
     public $changefreq;
     public $priority;
     public $image;
+
+    /**
+     * SitemapEntry constructor.
+     *
+     * @param null $location
+     * @param null $lastmod
+     * @param null $changefreq
+     * @param null $priority
+     * @param null $image
+     */
+    public function __construct($location = null, $lastmod = null, $changefreq = null, $priority = null, $image = null)
+    {
+        $this->location = $location;
+        $this->lastmod = $lastmod;
+        $this->changefreq = $changefreq;
+        $this->priority = $priority;
+        $this->image = $image;
+    }
 }